Adjust test suite to copy drive_manager.json during each run
Also disables installing the simulator.tgz (which included
drive_manager.json) as this is no longer necessary as the c7 openhpi
package now includes the simulator.

Note that this also disables installing of the lightning version of
zabbix-virtual-openhpi.  (This is now expected to be done before the
test is run.)

Change-Id: I92b0151245f4cd632effd164846f56a1ffcb2c70