Don't install packages during test suite
Instead, assumes packages are already installed.  This isn't necessarily
a good thing; arguably we should install any missing packages.  But as a
general rule, we want to avoid doing that while running the tests on the
jenkins drones and instead install everything up front.  So for now, we
can leave it off here too, and if it becomes an issue, we can provide an
additional script to install all the dependencies.

Change-Id: I9497e924b6d813e9e6e614b43f5a3a1ff8b4cbbd