rpm: fix build, was broken due to missing python36
A spec file must mention all the necessary dependencies, that are used
during the build, in the "BuildRequires" section, even if they're
already mentioned in the "Requires" section. Automation tools that set
up build environment rely on the correctness and completeness of this
information.

For example, `yum-builddep -y sspl-ll.spec` respects only
"BuildRequires" section. If it doesn't provide a full list of all the
build dependencies, a build is going to fail.

Example of such failure:

  error: Failed build dependencies:
         python36 is needed by systemd-python36-1.0.3-B65668_git1ce6c26.el7.x86_64