EOS-13291: Fix node hw sensor
Node hw sensor use last_sel_index file to generate alert for IPMI event.
Because of exception in processing some events, It was getting restarted(rescheduled)
again and again without updating last_sel_index. So it was sending alert for the events
(untill first exception in processing happens) in every run.
Because of this, NodeDatMsgHandler's message queue was getting flooded with messages and
It was not receiving new messages in queue.
Added error handling in Node Hw to avoid this issue

Signed-off-by: Sandeep Anjara <sandeep.anjara@seagate.com>